Adding and linking the new PrepExternalSchema version with the rest of the program.
Changes:
- Checks if the files given as input are FASTA files;
- Input files paths are divided into a number of lists equal to the
number of available cores. Input division is balanced based on the number
of sequences per file. This change reduces core idling;
- Users can redefine the BSR value (default=0.6);
- Include progress bar (still needs improvement);
- Ouputs file with the list of alleles that were determined to be invalid.
	Based on the following reasons:
		- ambiguous or invalid characters;
		- sequence length is not a multiple of 3;
		- inframe stop codon;
		- invalid start or stop codons.
- If a gene/locus file has no valid alleles, skips that case;
- Identifies alleles that code for the same protein sequence and chooses
only one of those alleles to BLAST and search for representatives, since
alleles with the same protein sequence have the same length and are considered
equal when searching for representatives (all alleles are included in the schema
file with all alleles);
- Starts by selecting the longest or one of the longest sequences as representative
to mimic the CreateSchema behaviour;
- Uses BLAST to search for representatives for each gene/locus until there is a
representative for every allele of that gene/locus. Keeps track of the alleles
that already have a representative and the number of alleles that still need to be
BLASTed, quickly reducing the number of BLAST searches and greatly reducing the
running time of the program.